requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
python-dotenv>=1.0.0